# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AI股票助手API",
    "description": "AI股票助手后端API接口文档",
    "version": "1.0.0",
    "contact": {
      "name": "AI股票助手开发团队"
    }
  },
  "host": "localhost:3033",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/stocks/sync": {
      "post": {
        "tags": ["股票管理"],
        "summary": "同步股票信息",
        "description": "通过API获取股票信息，不存在则新增，存在则更新",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SyncStockRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "同步成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/stocks/delete": {
      "post": {
        "tags": ["股票管理"],
        "summary": "删除股票",
        "description": "根据ID删除股票",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DeleteStockRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/stocks/list": {
      "post": {
        "tags": ["股票管理"],
        "summary": "获取股票列表",
        "description": "获取所有股票或根据条件筛选",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StockListRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/stocks/get-by-code": {
      "post": {
        "tags": ["股票管理"],
        "summary": "根据代码获取股票",
        "description": "根据股票代码获取股票信息",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetStockByCodeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/stocks/update-price": {
      "post": {
        "tags": ["股票管理"],
        "summary": "更新股票价格信息",
        "description": "更新股票的价格相关数据",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateStockPriceRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/stocks/update-holding": {
      "post": {
        "tags": ["股票管理"],
        "summary": "更新持仓信息",
        "description": "更新股票的持仓数量和成本信息",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateHoldingRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/stocks/calculate-market-value": {
      "post": {
        "tags": ["股票管理"],
        "summary": "计算市值",
        "description": "根据持仓数量和当前价格计算市值",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CalculateMarketValueRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "计算成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/stocks/batch-update": {
      "post": {
        "tags": ["股票管理"],
        "summary": "批量更新股票信息",
        "description": "批量更新多个股票的信息",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BatchUpdateStocksRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/stocks/update-pe": {
      "post": {
        "tags": ["股票管理"],
        "summary": "更新市盈率",
        "description": "更新指定股票的市盈率",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdatePERequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/stocks/update-volume": {
      "post": {
        "tags": ["股票管理"],
        "summary": "更新成交量",
        "description": "更新指定股票的成交量",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateVolumeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/trading/create": {
      "post": {
        "tags": ["交易记录"],
        "summary": "创建交易记录",
        "description": "创建单个交易记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateTradingRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "创建成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/trading/create-batch": {
      "post": {
        "tags": ["交易记录"],
        "summary": "批量创建交易记录",
        "description": "批量创建多个交易记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BatchCreateTradingRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "批量创建成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/trading/get-by-symbol": {
      "post": {
        "tags": ["交易记录"],
        "summary": "根据股票代码获取交易记录",
        "description": "根据股票代码获取相关交易记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetTradingBySymbolRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/trading/get-by-symbol-and-time": {
      "post": {
        "tags": ["交易记录"],
        "summary": "根据股票代码和时间范围获取交易记录",
        "description": "根据股票代码和时间范围获取相关交易记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetTradingBySymbolAndTimeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/trading/get-by-type": {
      "post": {
        "tags": ["交易记录"],
        "summary": "根据交易类型获取记录",
        "description": "根据买入或卖出类型获取交易记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetTradingByTypeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/trading/get-by-price-range": {
      "post": {
        "tags": ["交易记录"],
        "summary": "根据价格范围获取记录",
        "description": "根据价格范围获取交易记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetTradingByPriceRangeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/trading/get-latest": {
      "post": {
        "tags": ["交易记录"],
        "summary": "获取最新交易记录",
        "description": "获取最新的交易记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetLatestTradingRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/trading/stats": {
      "post": {
        "tags": ["交易记录"],
        "summary": "获取交易统计信息",
        "description": "获取交易相关的统计信息",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetTradingStatsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/trading/update": {
      "post": {
        "tags": ["交易记录"],
        "summary": "更新交易记录",
        "description": "更新指定的交易记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateTradingRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/trading/delete": {
      "post": {
        "tags": ["交易记录"],
        "summary": "删除交易记录",
        "description": "删除指定的交易记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DeleteTradingRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/trading/get-by-change-percent-range": {
      "post": {
        "tags": ["交易记录"],
        "summary": "根据涨跌幅范围获取记录",
        "description": "根据涨跌幅范围获取交易记录",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetTradingByChangePercentRangeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/trading/clean-old-data": {
      "post": {
        "tags": ["交易记录"],
        "summary": "清理过期数据",
        "description": "清理指定天数之前的过期交易数据",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CleanOldDataRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "清理成功",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "ApiResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "description": "返回的数据"
        },
        "message": {
          "type": "string",
          "description": "操作成功/报错信息"
        },
        "result": {
          "type": "integer",
          "description": "100表示接口正常，其他数字表示错误码"
        },
        "success": {
          "type": "boolean",
          "description": "true表示请求成功 false表示接口报错"
        }
      }
    },
...
    "Stock": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "description": "股票ID"
        },
        "code": {
          "type": "string",
          "description": "股票代码"
        },
        "name": {
          "type": "string",
          "description": "股票名称"
        },
        "market": {
          "type": "string",
          "description": "市场类型"
        },
        "marketCode": {
          "type": "integer",
          "description": "市场代码"
        },
        "pe": {
          "type": "number",
          "format": "decimal",
          "description": "市盈率"
        },
        "latestPrice": {
          "type": "number",
          "format": "decimal",
          "description": "最新价"
        },
        "changePercent": {
          "type": "number",
          "format": "decimal",
          "description": "涨跌幅(%)"
        },
        "changeAmount": {
          "type": "number",
          "format": "decimal",
          "description": "涨跌额"
        },
        "openPrice": {
          "type": "number",
          "format": "decimal",
          "description": "开盘价"
        },
        "highPrice": {
          "type": "number",
          "format": "decimal",
          "description": "最高价"
        },
        "lowPrice": {
          "type": "number",
          "format": "decimal",
          "description": "最低价"
        },
        "previousClosePrice": {
          "type": "number",
          "format": "decimal",
          "description": "昨收价"
        },
        "volume": {
          "type": "integer",
          "format": "int64",
          "description": "成交量(股)"
        },
        "holdingQuantity": {
          "type": "number",
          "format": "decimal",
          "description": "持仓数量"
        },
        "holdingCost": {
          "type": "number",
          "format": "decimal",
          "description": "持仓成本"
        },
        "marketValue": {
          "type": "number",
          "format": "decimal",
          "description": "市值"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间"
        }
      }
    },
...
    "SyncStockRequest": {
      "type": "object",
      "required": ["code", "marketCode"],
      "properties": {
        "code": {
          "type": "string",
          "description": "股票代码"
        },
        "marketCode": {
          "type": "integer",
          "description": "市场代码"
        }
      }
    },
    "DeleteStockRequest": {
      "type": "object",
      "required": ["id"],
      "properties": {
        "id": {
          "type": "integer",
          "description": "股票ID"
        }
      }
    },
    "StockListRequest": {
      "type": "object",
      "properties": {
        "market": {
          "type": "string",
          "description": "市场类型"
        },
        "marketCode": {
          "type": "integer",
          "description": "市场代码"
        },
        "page": {
          "type": "integer",
          "description": "页码"
        },
        "limit": {
          "type": "integer",
          "description": "每页数量"
        }
      }
    },
    "GetStockByCodeRequest": {
      "type": "object",
      "required": ["code"],
      "properties": {
        "code": {
          "type": "string",
          "description": "股票代码"
        }
      }
    },
    "UpdateStockPriceRequest": {
      "type": "object",
      "required": ["code"],
      "properties": {
        "code": {
          "type": "string",
          "description": "股票代码"
        },
        "latestPrice": {
          "type": "number",
          "format": "decimal",
          "description": "最新价"
        },
        "previousClosePrice": {
          "type": "number",
          "format": "decimal",
          "description": "昨收价"
        },
        "changePercent": {
          "type": "number",
          "format": "decimal",
          "description": "涨跌幅"
        },
        "changeAmount": {
          "type": "number",
          "format": "decimal",
          "description": "涨跌额"
        },
        "openPrice": {
          "type": "number",
          "format": "decimal",
          "description": "开盘价"
        },
        "highPrice": {
          "type": "number",
          "format": "decimal",
          "description": "最高价"
        },
        "lowPrice": {
          "type": "number",
          "format": "decimal",
          "description": "最低价"
        },
        "volume": {
          "type": "integer",
          "format": "int64",
          "description": "成交量"
        },
        "pe": {
          "type": "number",
          "format": "decimal",
          "description": "市盈率"
        }
      }
    },
    "UpdateHoldingRequest": {
      "type": "object",
      "required": ["code"],
      "properties": {
        "code": {
          "type": "string",
          "description": "股票代码"
        },
        "holdingQuantity": {
          "type": "number",
          "format": "decimal",
          "description": "持仓数量"
        },
        "holdingCost": {
          "type": "number",
          "format": "decimal",
          "description": "持仓成本"
        }
      }
    },
    "CalculateMarketValueRequest": {
      "type": "object",
      "required": ["code"],
      "properties": {
        "code": {
          "type": "string",
          "description": "股票代码"
        }
      }
    },
    "BatchUpdateStocksRequest": {
      "type": "object",
      "required": ["updates"],
      "properties": {
        "updates": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "股票代码"
              },
              "updateData": {
                "$ref": "#/definitions/Stock",
                "description": "更新数据"
              }
            }
          }
        }
      }
    },
    "UpdatePERequest": {
      "type": "object",
      "required": ["code", "pe"],
      "properties": {
        "code": {
          "type": "string",
          "description": "股票代码"
        },
        "pe": {
          "type": "number",
          "format": "decimal",
          "description": "市盈率"
        }
      }
    },
    "UpdateVolumeRequest": {
      "type": "object",
      "required": ["code", "volume"],
      "properties": {
        "code": {
          "type": "string",
          "description": "股票代码"
        },
        "volume": {
          "type": "integer",
          "format": "int64",
          "description": "成交量"
        }
      }
    },
...
    "CreateTradingRequest": {
      "type": "object",
      "properties": {
        "symbol": {
          "type": "string",
          "description": "股票代码"
        },
        "name": {
          "type": "string",
          "description": "股票名称"
        },
        "type": {
          "type": "string",
          "enum": ["buy", "sell"],
          "description": "交易类型"
        },
        "tradingTime": {
          "type": "string",
          "format": "date-time",
          "description": "成交时间"
        },
        "quantity": {
          "type": "integer",
          "format": "int64",
          "description": "交易数量"
        },
        "price": {
          "type": "number",
          "format": "decimal",
          "description": "成交价格"
        },
        "fee": {
          "type": "number",
          "format": "decimal",
          "description": "交易手续费"
        },
        "openPrice": {
          "type": "number",
          "format": "decimal",
          "description": "开盘价"
        },
        "changePercent": {
          "type": "number",
          "format": "decimal",
          "description": "涨跌幅"
        },
        "changeAmount": {
          "type": "number",
          "format": "decimal",
          "description": "涨跌额"
        },
        "remarks": {
          "type": "string",
          "description": "备注"
        }
      }
    },
    "BatchCreateTradingRequest": {
      "type": "object",
      "required": ["tradingDataList"],
      "properties": {
        "tradingDataList": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CreateTradingRequest"
          }
        }
      }
    },
    "GetTradingBySymbolRequest": {
      "type": "object",
      "required": ["symbol"],
      "properties": {
        "symbol": {
          "type": "string",
          "description": "股票代码"
        }
      }
    },
    "GetTradingBySymbolAndTimeRequest": {
      "type": "object",
      "required": ["symbol", "startTime", "endTime"],
      "properties": {
        "symbol": {
          "type": "string",
          "description": "股票代码"
        },
        "startTime": {
          "type": "string",
          "format": "date-time",
          "description": "开始时间"
        },
        "endTime": {
          "type": "string",
          "format": "date-time",
          "description": "结束时间"
        }
      }
    },
    "GetTradingByTypeRequest": {
      "type": "object",
      "required": ["type"],
      "properties": {
        "type": {
          "type": "string",
          "enum": ["buy", "sell"],
          "description": "交易类型"
        }
      }
    },
    "GetTradingByPriceRangeRequest": {
      "type": "object",
      "required": ["minPrice", "maxPrice"],
      "properties": {
        "minPrice": {
          "type": "number",
          "format": "decimal",
          "description": "最低价格"
        },
        "maxPrice": {
          "type": "number",
          "format": "decimal",
          "description": "最高价格"
        }
      }
    },
    "GetLatestTradingRequest": {
      "type": "object",
      "properties": {
        "symbol": {
          "type": "string",
          "description": "股票代码"
        },
        "limit": {
          "type": "integer",
          "description": "限制数量"
        }
      }
    },
    "GetTradingStatsRequest": {
      "type": "object",
      "properties": {
        "symbol": {
          "type": "string",
          "description": "股票代码"
        },
        "startTime": {
          "type": "string",
          "format": "date-time",
          "description": "开始时间"
        },
        "endTime": {
          "type": "string",
          "format": "date-time",
          "description": "结束时间"
        }
      }
    },
    "UpdateTradingRequest": {
      "type": "object",
      "required": ["id", "updateData"],
      "properties": {
        "id": {
          "type": "integer",
          "description": "交易记录ID"
        },
        "updateData": {
          "$ref": "#/definitions/CreateTradingRequest",
          "description": "更新数据"
        }
      }
    },
    "DeleteTradingRequest": {
      "type": "object",
      "required": ["id"],
      "properties": {
        "id": {
          "type": "integer",
          "description": "交易记录ID"
        }
      }
    },
    "GetTradingByChangePercentRangeRequest": {
      "type": "object",
      "required": ["minChange", "maxChange"],
      "properties": {
        "minChange": {
          "type": "number",
          "format": "decimal",
          "description": "最小涨跌幅"
        },
        "maxChange": {
          "type": "number",
          "format": "decimal",
          "description": "最大涨跌幅"
        }
      }
    },
    "CleanOldDataRequest": {
      "type": "object",
      "properties": {
        "daysToKeep": {
          "type": "integer",
          "description": "保留天数"
        }
      }
    },
...
  },
...
}